- introduce new "capability" method for SQL::Statement and SQL::Eval::Table - expect every table object being derived from SQL::Eval::Table - rewrite DELETE and UPDATE command based on table capabilities - add abstract methods for all methods derived classes must override   (this means, open_table for SQL::Statement deriveds must be overridden    and all data access methods of tables - see SQL::Eval::Table for details) - Tests are fixed to use TEMP TABLES explicitely when required - Method documentation of SQL::Statement and SQL::Eval::Table are improved
git-svn-id: https://svn.perl.org/modules/SQL-Statement/trunk@13970 50811bd7-b8ce-0310-adc1-d9db26280581

+++ lib/SQL/Statement.pm
@@ -64,7 +64,8 @@ sub new
 
         $parser = SQL::Parser->new( $parser_dialect, $flags );
     }
-    $self->{termFactory} = SQL::Statement::TermFactory->new($self);
+    $self->{termFactory}  = SQL::Statement::TermFactory->new($self);
+    $self->{capabilities} = {};
     $self->prepare( $sql, $parser );
     return $self;
 }
@@ -220,15 +221,17 @@ sub CALL
 sub DROP ($$$)
 {
     my ( $self, $data, $params ) = @_;
+    my $eval;
+    eval { ($eval) = $self->open_tables( $data, 0, 1 ); };
     if ( $self->{ignore_missing_table} )
     {
-        eval { $self->open_tables( $data, 0, 0 ) };
         if ( $@ and $@ =~ m/no such (table|file)/i )
         {
             return ( -1, 0 );
         }
     }
-    my ($eval) = $self->open_tables( $data, 0, 1 );
+
+    $self->do_err($@) if ($@);
 
     #    return undef unless $eval;
     return ( -1, 0 ) unless $eval;
@@ -308,44 +311,58 @@ sub DELETE ($$$)
     return unless $eval;
     $eval->params($params);
     $self->verify_columns( $data, $eval, $all_cols );
-    my ($table)    = $eval->table( $self->tables(0)->name() );
+    my $tname      = $self->tables(0)->name();
+    my ($table)    = $eval->table($tname);
     my ($affected) = 0;
     my ( @rows, $array );
 
-    if ( $table->can('delete_one_row') )
+    while ( $array = $table->fetch_row($data) )
     {
-        while ( my $array = $table->fetch_row($data) )
+        if ( $self->eval_where( $eval, $tname, $array ) )
         {
-            if ( $self->eval_where( $eval, '', $array ) )
+            ++$affected;
+            if ( $table->capability('rowwise_delete') and $table->capability('inplace_delete') )
             {
-                ++$affected;
-                $array = $self->{fetched_value} if ( $self->{fetched_from_key} );
-                $table->delete_one_row( $data, $array );
-                return ( $affected, 0 ) if ( $self->{fetched_from_key} );
+                if ( $table->capability('delete_one_row') )
+                {
+                    $table->delete_one_row( $data, $array );
+                }
+                elsif ( $table->capability('delete_current_row') )
+                {
+                    $table->delete_current_row( $data, $array );
+                }
+            }
+            elsif ( $table->capability('rowwise_delete') )
+            {
+                push( @rows, $array );
             }
         }
 
-        return ( $affected, 0 );
+        push( @rows, $array ) unless ( $table->capability('rowwise_delete') );
     }
 
-    while ( $array = $table->fetch_row($data) )
+    if (@rows)
     {
-        if ( $self->eval_where( $eval, '', $array ) )
+        if ( $table->capability('rowwise_delete') )    # @rows is empty in case of inplace_delete capability
         {
-            ++$affected;
+            foreach my $array (@rows)
+            {
+                $table->delete_one_row( $data, $array );
+            }
         }
         else
         {
-            push( @rows, $array );
+            # rewrite table without deleted elements
+            $table->seek( $data, 0, 0 );
+            foreach my $array (@rows)
+            {
+                $table->push_row( $data, $array );
+            }
+            $table->truncate($data);
         }
     }
-    $table->seek( $data, 0, 0 );
-    foreach $array (@rows)
-    {
-        $table->push_row( $data, $array );
-    }
-    $table->truncate($data);
-    ( $affected, 0 );
+
+    return ( $affected, 0 );
 }
 
 sub UPDATE ($$$)
@@ -368,20 +385,17 @@ sub UPDATE ($$$)
 
     while ( my $array = $table->fetch_row($data) )
     {
+        my $originalValues;
         if ( $self->eval_where( $eval, $tname, $array ) )
         {
             my $valpos = 0;
-            my $originalValues;
-            if ( not( $self->{fetched_from_key} )
-                 and $table->can('update_specific_row') )
+            if ( $table->capability('update_specific_row') )
             {
                 $originalValues = clone($array);
-                $array          = $self->{fetched_value};
             }
 
             for ( my $i = 0; $i < $self->columns(); $i++ )
             {
-                my $col = $self->columns($i);
                 my $val = $self->row_values( 0, $i );
                 if ( defined( _INSTANCE( $val, 'SQL::Statement::Param' ) ) )
                 {
@@ -405,44 +419,65 @@ sub UPDATE ($$$)
                     return $self->do_err('Internal error: Unexpected column type');
                 }
 
+                my $col = $self->columns($i);
                 $array->[ $table->column_num( $col->name() ) ] = $val;
             }
 
             ++$affected;
-
-            # Martin Fabiani <martin@fabiani.net>:
-            # the following block is the most important enhancement to SQL::Statement::UPDATE
-            if ( !$self->{fetched_from_key} )
+            if ( $table->capability('rowwise_update') and $table->capability('inplace_update') )
             {
-                if ( $table->can('update_specific_row') )
+                # Martin Fabiani <martin@fabiani.net>:
+                # the following block is the most important enhancement to SQL::Statement::UPDATE
+                if ( $table->capability('update_specific_row') )
                 {
                     $table->update_specific_row( $data, $array, $originalValues );
-                    next;
                 }
-                elsif ( $table->can('update_one_row') )
+                elsif ( $table->capability('update_one_row') )
                 {
+                    # NOTE: this prevents from updating index keys
                     $table->update_one_row( $data, $array );
-                    next;
+                }
+                elsif ( $table->capability('update_current_row') )
+                {
+                    $table->update_current_row( $data, $array );
                 }
             }
-            else
+            elsif ( $table->capability('rowwise_update') )
             {
-                $table->update_one_row( $data, $array );
-                return ( $affected, 0 );
+                push( @rows, $array ) unless ( $table->capability('update_specific_row') );
+                push( @rows, [ $array, $originalValues ] ) if ( $table->capability('update_specific_row') );
             }
         }
 
-        push( @rows, $array ) unless ( $table->can('update_one_row') || $table->can('update_specific_row') );
+        push( @rows, $array ) unless ( $table->capability('rowwise_update') );
     }
 
-    unless ( $table->can('update_one_row') || $table->can('update_specific_row') )
+    if (@rows)
     {
-        $table->seek( $data, 0, 0 );
-        foreach my $array (@rows)
+        if ( $table->capability('rowwise_update') )    # @rows is empty in case of inplace_update capability
         {
-            $table->push_row( $data, $array );
+            foreach my $array (@rows)
+            {
+                if ( $table->capability('update_specific_row') )
+                {
+                    $table->update_specific_row( $data, $array->[0], $array->[1] );
+                }
+                elsif ( $table->capability('update_one_row') )
+                {
+                    $table->update_one_row( $data, $array );
+                }
+            }
         }
-        $table->truncate($data);
+        else
+        {
+            # rewrite table with updated elements
+            $table->seek( $data, 0, 0 );
+            foreach my $array (@rows)
+            {
+                $table->push_row( $data, $array );
+            }
+            $table->truncate($data);
+        }
     }
 
     return ( $affected, 0 );
@@ -956,16 +991,13 @@ sub SELECT($$)
         {
             next if ( defined($limit_count) and ( $row_count++ < $offset ) );
             $limit_count++ if ( defined($limit_count) );
-            $array = $self->{fetched_value} if ( $self->{fetched_from_key} );
 
             my @row = map { $_->value($e) } $self->columns();
             push( @{$rows}, \@row );
 
-            # We quit here if its a primary key search
-            # or if there's a limit clause without order clause
+            # We quit here if there's a limit clause without order clause
             # and the limit has been reached
-            #
-            if ( $self->{fetched_from_key} || ( defined($limit_count) && ( $limit_count >= $self->limit() ) ) )
+            if ( defined($limit_count) && ( $limit_count >= $self->limit() ) )
             {
                 return ( scalar( @{$rows} ), $numFields, $rows );
             }
@@ -975,9 +1007,10 @@ sub SELECT($$)
     if ( $self->distinct() )
     {
         my %seen;
-        @{$rows} = map
-        {
-            $seen{ join( "\0", ( map { defined($_) ? $_ : '' } @{$_} ) ) }++ ? () : $_
+        @{$rows} = map {
+            $seen{ join( "\0", ( map { defined($_) ? $_ : '' } @{$_} ) ) }++
+              ? ()
+              : $_
         } @{$rows};
     }
 
@@ -1078,6 +1111,8 @@ sub fetch
     return $row;
 }
 
+sub open_table ($$$$$) { croak "Abstract method " . ref( $_[0] ) . "::open_table called" }
+
 sub open_tables
 {
     my ( $self, $data, $createMode, $lockMode ) = @_;
@@ -1110,12 +1145,13 @@ sub open_tables
         elsif ( $data->{Database}->{sql_ram_tables}->{$name} )
         {
             $t->{$name} = $data->{Database}->{sql_ram_tables}->{$name};
-            $t->{$name}->{index} = 0;
+            $t->{$name}->seek( $data, 0, 0 );
             $t->{$name}->init_table( $data, $name, $createMode, $lockMode ) if ( $t->{$name}->can('init_table') );
         }
-        elsif ( $self->{is_ram_table} or !( $self->can('open_table') ) )
+        elsif ( $self->{is_ram_table} )
         {
-            $t->{$name} = $data->{Database}->{sql_ram_tables}->{$name} = SQL::Statement::RAM->new( $name, [], [] );
+            $t->{$name} = $data->{Database}->{sql_ram_tables}->{$name} =
+              SQL::Statement::RAM::Table->new( $name, [], [] );
         }
         else
         {
@@ -1806,9 +1842,17 @@ sub get_user_func_table
     # my $tempTable = SQL::Statement::TempTable->new(
     #     $name, $col_names, $col_names, $data_aryref
     # );
-    my $tempTable = SQL::Statement::RAM->new( $name, $col_names, $data_aryref );
+    my $tempTable = SQL::Statement::RAM::Table->new( $name, $col_names, $data_aryref );
     $tempTable->{all_cols} ||= $col_names;
     return $tempTable;
+}
+
+sub capability($)
+{
+    my ( $self, $capname ) = @_;
+    return $self->{capabilities}->{$capname} if ( defined( $self->{capabilities}->{$capname} ) );
+
+    return;
 }
 
 sub DESTROY
@@ -1826,8 +1870,6 @@ sub DESTROY
     undef $self->{computed_column};
     undef $self->{cur_table};
     undef $self->{data};
-    undef $self->{fetched_value};
-    undef $self->{fetched_from_key};
     undef $self->{group_by};
     undef $self->{has_OR};
     undef $self->{join};
@@ -2035,9 +2077,13 @@ sub getAffectedResult    # (\@)
 package SQL::Statement::TempTable;
 
 use vars qw(@ISA);
-require SQL::Eval;
 
-@ISA = qw(SQL::Eval::Table);
+BEGIN
+{
+    require SQL::Eval;
+
+    @SQL::Statement::TempTable::ISA = qw(SQL::Eval::Table);
+}
 
 sub new
 {
@@ -2058,7 +2104,7 @@ sub new
                  rowpos     => 0,
                  maxrow     => scalar @$table
                };
-    return bless $self, $class;
+    return $class->SUPER::new($self);
 }
 
 sub is_shared($) { $_[0]->{is_shared}->{ $_[1] }; }
@@ -2068,14 +2114,14 @@ sub column_num($)
 {
     my ( $s, $col ) = @_;
     my $new_col = $s->{col_nums}->{$col};
-    if ( !defined $new_col )
+    unless ( defined($new_col) )
     {
-        my @tmp = split '~', $col;
-        return undef unless ( 2 == scalar(@tmp) );
+        my @tmp = split( '~', $col );
+        return unless ( 2 == scalar(@tmp) );
         $new_col = lc( $tmp[0] ) . '~' . $tmp[1];
         $new_col = $s->{col_nums}->{$new_col};
     }
-    $new_col;
+    return $new_col;
 }
 
 sub fetch_row()
@@ -2236,9 +2282,113 @@ directions. If you want to join the development team, or just hear more
 about the development, write Jeff (<jzuckerATcpan.org>) or Jens
 (<rehsackATcpan.org>) a note.
 
-=begin undocumented
-
 =head1 METHODS
+
+Following methods can or must be overriden by derived classes.
+
+=head2 capability
+
+This method is called for quicker check for capabilies than
+C<< $self->can('method_name') >>. Currently no capabilities for the Statement
+objects are required - but analogous to C<< SQL::Eval::Table::capability >>
+it's declared for future use.
+
+=head2 open_table
+
+The C<< open_table >> method must be overriden by derived classes to provide
+the capability of opening data tables. This is is must have.
+
+Arguments given to open_table call:
+
+=over 4
+
+=item C<< $data >>
+
+The database memo parameter. See L</execute>.
+
+=item C<< $table >>
+
+Name of the table to open as parsed from SQL statement.
+
+=item C<< $createMode >>
+
+Flag if the table shall opened in create mode (C<< CREATE TABLE ... >>) or
+any other one. Set to true value in create mode.
+
+=item C<< $lockMode >>
+
+Flag if the table shall opened for writing (any other than C<< SELECT ... >>).
+Set to a true value is this case.
+
+=back
+
+Following methods are required to use SQL::Statement in a DBD (for example).
+
+=head2 new
+
+Instantiates a new SQL::Statement object
+
+Arguments:
+
+=over 4
+
+=item C<< $sql >>
+
+The SQL statement for later actions
+
+=item C<< $parser >>
+
+Instance of a L<SQL::Parser> object or flags for it's instantiation.
+If omitted, default flags are used.
+
+=back
+
+When the basic initalization is completed,
+C<< $self->prepare($sql, $parser) >> is invoked.
+
+=head2 prepare
+
+Prepares SQL::Statement to execute a SQL statement.
+
+Arguments:
+
+=over 4
+
+=item C<< $sql >>
+
+The SQL statement to parse and prepare
+
+=item C<< $parser >>
+
+Instance of a L<SQL::Parser> object to parse given SQL statement.
+
+=back
+
+=head2 execute
+
+Executes prepared statement.
+
+Arguments:
+
+=over 4
+
+=item C<< $data >>
+
+Memo field passed through to calls of instantiated C<< $table >> objects or
+C<< open_table >> calls. In C<< CREATE >> with subquery,
+C<< $data->{Database} >> must be a DBI database handle object.
+
+=item C<< $params >>
+
+Bound params via DBI ...
+
+=back
+
+=head2 errstr
+
+Gives the error string of the last error, if any.
+
+=begin undocumented
 
 =head2 _anycmp
 
@@ -2260,11 +2410,7 @@ about the development, write Jeff (<jzuckerATcpan.org>) or Jens
 
 =head2 do_err
 
-=head2 errstr
-
 =head2 eval_where
-
-=head2 execute
 
 =head2 fetch
 
@@ -2290,8 +2436,6 @@ about the development, write Jeff (<jzuckerATcpan.org>) or Jens
 
 =head2 params
 
-=head2 prepare
-
 =head2 row_values
 
 =head2 run_functions
@@ -2310,7 +2454,35 @@ about the development, write Jeff (<jzuckerATcpan.org>) or Jens
 
 =end undocumented
 
-=head1 Where can I go for more help?
+=head1 SUPPORT
+
+You can find documentation for this module with the perldoc command.
+
+    perldoc SQL::Statement
+
+You can also look for information at:
+
+=over 4
+
+=item * RT: CPAN's request tracker
+
+L<http://rt.cpan.org/NoAuth/Bugs.html?Dist=SQL-Statement>
+
+=item * AnnoCPAN: Annotated CPAN documentation
+
+L<http://annocpan.org/dist/SQL-Statement>
+
+=item * CPAN Ratings
+
+L<http://cpanratings.perl.org/s/SQL-Statement>
+
+=item * Search CPAN
+
+L<http://search.cpan.org/dist/SQL-Statement/>
+
+=back
+
+=head2 Where can I go for more help?
 
 For questions about installation or usage, please ask on the
 dbi-users@perl.org mailing list or post a question on PerlMonks
@@ -2431,6 +2603,11 @@ of C<SQL::Statement>, L<DBD::File> and L<DBD::CSV> and must be a general
 improvement.
 
 =head1 AUTHOR AND COPYRIGHT
+
+Jochen Wiedmann created the original module as an XS (C) extension in 1998.
+Jeff Zucker took over the maintenance in 2001 and rewrote all of the C
+portions in perl and began extending the SQL support. Since 2008, Jens
+Rehsack holds the maintainership.
 
 Copyright (c) 2001,2005 by Jeff Zucker: jzuckerATcpan.org
 Copyright (c) 2007-2010 by Jens Rehsack: rehsackATcpan.org
